server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
elevenlabs==1.51.0
fastapi==0.115.8
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a2
pydantic_core==2.29.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
